Refactor specifiers to components
