+++ flask_swagger_generator/components/__init__.py
@@ -1,0 +1,3 @@
+from flask_swagger_generator.components.swagger_version import SwaggerVersion
+
+__all__ = ['SwaggerVersion']
