+++ flask_swagger_generator/components/version_three/__init__.py
@@ -1,0 +1,45 @@
+from flask_swagger_generator.components.version_three.operation_id import \
+    SwaggerOperationId
+from flask_swagger_generator.components.version_three.parameters import \
+    SwaggerParameters
+from flask_swagger_generator.components.version_three.path import SwaggerPath
+from flask_swagger_generator.components.version_three.path_parameter import \
+    SwaggerPathParameter
+from flask_swagger_generator.components.version_three.query_parameter import \
+    SwaggerQueryParameter
+from flask_swagger_generator.components.version_three.request_body import \
+    SwaggerRequestBody, SwaggerRequestBodyComponent
+from flask_swagger_generator.components.version_three.request_type import \
+    SwaggerRequestType
+from flask_swagger_generator.components.version_three.responses import \
+    SwaggerResponses
+from flask_swagger_generator.components.version_three.response import \
+    SwaggerResponse, SwaggerResponseComponent
+from flask_swagger_generator.components.version_three.schema_attribute import \
+    SwaggerSchemaAttribute
+from flask_swagger_generator.components.version_three.schemas import \
+    SwaggerSchema, SwaggerMashmallowSchema
+from flask_swagger_generator.components.version_three.security import \
+    SwaggerSecurity, SwaggerSecurityComponent
+from flask_swagger_generator.components.version_three.tag import SwaggerTag
+
+
+__all__ = [
+    "SwaggerOperationId",
+    "SwaggerParameters",
+    "SwaggerPath",
+    "SwaggerPathParameter",
+    "SwaggerQueryParameter",
+    "SwaggerRequestBody",
+    "SwaggerRequestType",
+    "SwaggerResponses",
+    "SwaggerResponse",
+    "SwaggerSchema",
+    "SwaggerSchemaAttribute",
+    "SwaggerSecurity",
+    "SwaggerTag",
+    "SwaggerMashmallowSchema",
+    "SwaggerRequestBodyComponent",
+    "SwaggerResponseComponent",
+    "SwaggerSecurityComponent"
+]
