+++ flask_swagger_generator/utils/__init__.py
@@ -3,11 +3,17 @@ from .security_type import SecurityType
 from .schema_type import SchemaType
 from .input_type import InputType
 from .request_type import RequestType
+from .random import random_string
+from .swagger_version import SwaggerVersion
+from .property_type import PropertyType
 
 __all__ = [
     'ParameterType',
     'SecurityType',
     'SchemaType',
     'InputType',
-    'RequestType'
+    'RequestType',
+    "SwaggerVersion",
+    "random_string",
+    "PropertyType"
 ]
