+++ tests/resources/swagger_version_three/all_components_swagger.yaml
@@ -1,0 +1,268 @@
+openapi: 3.0.1
+info:
+  title: Application
+  description: Generated at 24/10/2022 18:19:41. This is the swagger 
+    ui based on the open api 3.0 specification of the Application
+  version: 1.0.0 created by the flask swagger generator.
+externalDocs:
+  description: Find out more about Swagger
+  url: 'http://swagger.io'
+servers:
+  - url: /
+paths:
+  '/objects':
+    get:
+      tags:
+      - objects
+      operationId: 'list_objects'
+      parameters:
+      - in: query
+        name: param1
+        schema:
+          type: string
+        description: None
+        required: False
+        allowReserved: False
+      - in: query
+        name: param2
+        schema:
+          type: integer
+        description: None
+        required: False
+        allowReserved: False
+      responses:
+        '200':
+          $ref: '#/components/responses/list_objects_response'
+  '/objects/{object_id}':
+    get:
+      tags:
+      - objects
+      operationId: 'retrieve_object'
+      parameters:
+      - in: path
+        name: object_id
+        schema:
+          type: integer
+        description: None
+        required: True
+      responses:
+        '200':
+          $ref: '#/components/responses/retrieve_object_response'
+    delete:
+      tags:
+      - objects
+      operationId: 'delete_object'
+      parameters:
+      - in: path
+        name: object_id
+        schema:
+          type: integer
+        description: None
+        required: True
+      responses:
+        '204':
+          $ref: '#/components/responses/delete_object_response'
+      security:
+        - bearerAuth: []
+    put:
+      tags:
+      - objects
+      operationId: 'update_object'
+      parameters:
+      - in: path
+        name: object_id
+        schema:
+          type: integer
+        description: None
+        required: True
+      requestBody:
+        $ref: '#/components/requestBodies/update_object_request_body'
+      responses:
+        '200':
+          $ref: '#/components/responses/update_object_response'
+      security:
+        - bearerAuth: []
+    post:
+      tags:
+      - objects
+      operationId: 'create_object'
+      parameters:
+      - in: path
+        name: object_id
+        schema:
+          type: integer
+        description: None
+        required: True
+      requestBody:
+        $ref: '#/components/requestBodies/create_object_request_body'
+      responses:
+        '201':
+          $ref: '#/components/responses/create_object_response'
+components:
+  securitySchemes:
+    bearerAuth:
+      type: http
+      scheme: bearer
+      bearerFormat: JWT
+  requestBodies:
+    update_object_request_body:
+      description: None
+      required: True
+      content:
+         application/json:
+             schema:
+                 $ref: '#/components/schemas/update_object'
+    create_object_request_body:
+      description: None
+      required: True
+      content:
+         application/json:
+             schema:
+                 $ref: '#/components/schemas/create_object'
+  responses:
+    list_objects_response:
+      description: list_objects response
+      content:
+         application/json:
+             schema:
+                 $ref: '#/components/schemas/list_objects_response_schema'
+    retrieve_object_response:
+      description: retrieve_object response
+      content:
+         application/json:
+             schema:
+                 $ref: '#/components/schemas/retrieve_object_response_schema'
+    delete_object_response:
+      description: delete_object response
+      content:
+         application/json:
+             schema:
+                 $ref: '#/components/schemas/delete_object_response_schema'
+    update_object_response:
+      description: update_object response
+      content:
+         application/json:
+             schema:
+                 $ref: '#/components/schemas/update_object_response_schema'
+    create_object_response:
+      description: create_object response
+      content:
+         application/json:
+             schema:
+                 $ref: '#/components/schemas/create_object_response_schema'
+  schemas:
+    list_objects_response_schema:
+      type: object
+      properties:
+        id:
+          type: integer
+          example: 10
+        name:
+          type: string
+          example: level01
+        nested_dict:
+          $ref: "#/components/schemas/nested_dict_list_objects_response_schema"
+        nested_list:
+          type: array
+          items:
+            type: object
+    nested_dict_list_objects_response_schema:
+      type: object
+      properties:
+        id:
+          type: integer
+          example: 1
+        name:
+          type: string
+          example: level02
+        values:
+          $ref: "#/components/schemas/values_nested_dict_list_objects_response_schema"
+    values_nested_dict_list_objects_response_schema:
+      type: object
+      properties:
+        id:
+          type: integer
+          example: 3
+        name:
+          type: string
+          example: level03
+    retrieve_object_response_schema:
+      type: object
+      properties:
+        id:
+          type: integer
+          example: 10
+        name:
+          type: string
+          example: test_object
+    delete_object_response_schema:
+      type: object
+      properties:
+        id:
+          type: integer
+          example: 20
+        name:
+          type: string
+          example: test_object
+    update_object:
+      type: object
+      properties:
+        id:
+          type: integer
+          example: 10
+        name:
+          type: string
+          example: test_object
+    update_object_response_schema:
+      type: object
+      properties:
+        id:
+          type: integer
+          example: 10
+        name:
+          type: string
+          example: level01
+        nested_dict:
+          $ref: "#/components/schemas/nested_dict_update_object_response_schema"
+        nested_list:
+          type: array
+          items:
+            type: object
+    nested_dict_update_object_response_schema:
+      type: object
+      properties:
+        id:
+          type: integer
+          example: 1
+        name:
+          type: string
+          example: level02
+        values:
+          $ref: "#/components/schemas/values_nested_dict_update_object_response_schema"
+    values_nested_dict_update_object_response_schema:
+      type: object
+      properties:
+        id:
+          type: integer
+          example: 3
+        name:
+          type: string
+          example: level03
+    create_object:
+      type: object
+      properties:
+        id:
+          type: integer
+          example: 10
+        name:
+          type: string
+          example: test_object
+    create_object_response_schema:
+      type: object
+      properties:
+        id:
+          type: integer
+          example: 10
+        name:
+          type: string
+          example: test_object
