+++ tests/resources/swagger_version_three/query_params_swagger.yaml
@@ -1,0 +1,46 @@
+openapi: 3.0.1
+info:
+  title: Application
+  description: Generated at 24/10/2022 18:20:06. This is the swagger 
+    ui based on the open api 3.0 specification of the Application
+  version: 1.0.0 created by the flask swagger generator.
+externalDocs:
+  description: Find out more about Swagger
+  url: 'http://swagger.io'
+servers:
+  - url: /
+paths:
+  '/objects':
+    get:
+      tags:
+      - objects
+      operationId: 'list_objects'
+      parameters:
+      - in: query
+        name: param1
+        schema:
+          type: string
+        description: None
+        required: False
+        allowReserved: False
+      - in: query
+        name: param2
+        schema:
+          type: integer
+        description: None
+        required: False
+        allowReserved: False
+      responses:
+        '200':
+          $ref: '#/components/responses/list_objects_response'
+components:
+  responses:
+    list_objects_response:
+      description: list_objects response
+      content:
+         application/json:
+             schema:
+                 $ref: '#/components/schemas/list_objects_response_schema'
+  schemas:
+    list_objects_response_schema:
+      type: object
