+++ tests/swagger_version_three/test_all_components.py
@@ -1,0 +1,109 @@
+import os
+import pathlib
+
+from deepdiff import DeepDiff
+from flask import Blueprint, jsonify
+from flask_testing import TestCase
+
+from flask_swagger_generator.generators import Generator
+from flask_swagger_generator.components import SwaggerVersion
+from flask_swagger_generator.utils import SecurityType
+from tests.resources.app_base import create_app
+from tests.resources.utils import yaml_as_dict
+
+app = create_app()
+generator = Generator.of(SwaggerVersion.VERSION_THREE)
+blueprint = Blueprint('objects', __name__)
+query_params = [
+    {
+        "name": "param1",
+        "type": "string",
+    },
+    {
+        "name": "param2",
+        "type": "int",
+    }
+]
+schema_one = {
+    "id": 10,
+    "name": "level01",
+    "nested_dict": {
+        "id":1,
+        "name":"level02",
+        "values" : {
+            "id": 3,
+            "name": "level03"
+        }
+    },
+    "nested_list": [
+        {"a":1},
+        {"b":2}
+    ]
+}
+schema_two = {'id': 10, 'name': 'test_object'}
+schema_three = {'id': 20, 'name': 'test_object'}
+
+
+@generator.response(200, schema_one)
+@generator.query_parameters(query_params)
+@blueprint.route('/objects', methods=['GET'])
+def list_objects():
+    return jsonify({'objects': []}), 200
+
+
+@generator.response(200, schema_two)
+@blueprint.route('/objects/<int:object_id>', methods=['GET'])
+def retrieve_object(object_id, child_id):
+    return jsonify({'objects': []}), 200
+
+
+@generator.response(204, schema_three)
+@generator.security(SecurityType.BEARER_AUTH)
+@blueprint.route('/objects/<int:object_id>', methods=['DELETE'])
+def delete_object(object_id):
+    return jsonify({'objects': []}), 200
+
+
+@generator.response(200, schema_one)
+@generator.security(SecurityType.BEARER_AUTH)
+@generator.request_body(schema_two)
+@blueprint.route('/objects/<int:object_id>', methods=['PUT'])
+def update_object(object_id):
+    pass
+
+
+@generator.response(status_code=201, schema={'id': 10, 'name': 'test_object'})
+@generator.request_body({'id': 10, 'name': 'test_object'})
+@blueprint.route('/objects/<int:object_id>', methods=['POST'])
+def create_object(object_id):
+    return jsonify({'objects': []}), 200
+
+
+app.register_blueprint(blueprint)
+
+
+class AppTestBase(TestCase):
+
+    def setUp(self) -> None:
+        super(AppTestBase, self).setUp()
+
+    def create_app(self):
+        return app
+
+    def tearDown(self):
+        super(AppTestBase, self).tearDown()
+
+    def test(self):
+        generated_swagger_path = os.path.join(
+            pathlib.Path(__file__).parent.parent.absolute()
+        ) + '/resources/generated.yaml'
+        correct_swagger_path = os.path.join(
+            pathlib.Path(__file__).parent.parent.absolute()
+        ) + '/resources/swagger_version_three/all_components_swagger.yaml'
+        generator.generate_swagger(self.app, generated_swagger_path)
+        generated_yaml = yaml_as_dict(generated_swagger_path)
+        reference_yaml = yaml_as_dict(correct_swagger_path)
+        reference_yaml['info'].pop('description')
+        generated_yaml['info'].pop('description')
+        difference = DeepDiff(generated_yaml, reference_yaml, ignore_order=True)
+        self.assertEqual({}, difference)
