+++ tests/swagger_version_three/test_marshmallow_schema_component.py
@@ -1,20 +1,18 @@
 import os
-from deepdiff import DeepDiff
 import pathlib
 
-from flask_testing import TestCase
-from tests.resources.test_app import create_app
+from deepdiff import DeepDiff
 from flask import Blueprint, jsonify
-from flask_swagger_generator.generators import Generator
-from flask_swagger_generator.specifiers import SwaggerVersion
-from flask_swagger_generator.utils import SecurityType
-
+from flask_testing import TestCase
 from marshmallow import Schema, fields
 
+from flask_swagger_generator.generators import Generator
+from flask_swagger_generator.components import SwaggerVersion
+from flask_swagger_generator.utils import SecurityType
+from tests.resources.app_base import create_app
 from tests.resources.utils import yaml_as_dict
 
 app = create_app()
-
 generator = Generator.of(SwaggerVersion.VERSION_THREE)
 blueprint = Blueprint('objects', __name__)
 
@@ -38,31 +36,44 @@ class ObjectDeserializer(Schema):
     attribute_six = fields.Nested(ObjectChildDeserializer(many=False))
 
 
-schema_two = generator.create_schema(
-    'schema_two', {'id': 10, 'name': 'test_object'}
-)
-schema_three = generator.create_schema('schema_three', ObjectDeserializer())
+query_params = [
+    {
+        "name": "param1",
+        "type": "string",
+    },
+    {
+        "name": "param2",
+        "type": "int",
+    }
+]
 
 
-@generator.response(200, schema_two)
+@generator.response(200, ObjectSerializer())
+@generator.query_parameters(query_params)
+@blueprint.route('/objects', methods=['GET'])
+def list_objects():
+    return jsonify({'objects': []}), 200
+
+
+@generator.response(200, ObjectSerializer())
 @blueprint.route('/objects/<int:object_id>', methods=['GET'])
 def retrieve_object(object_id, child_id):
     return jsonify({'objects': []}), 200
 
 
-@generator.response(204, schema_two)
+@generator.response(204, ObjectSerializer())
 @generator.security(SecurityType.BEARER_AUTH)
 @blueprint.route('/objects/<int:object_id>', methods=['DELETE'])
 def delete_object(object_id):
     return jsonify({'objects': []}), 200
 
 
-@generator.response(200, schema_two)
+@generator.response(200, ObjectSerializer())
 @generator.security(SecurityType.BEARER_AUTH)
-@generator.request_body(schema_three)
+@generator.request_body(ObjectDeserializer())
 @blueprint.route('/objects/<int:object_id>', methods=['PUT'])
 def update_object(object_id):
-    pass
+    return jsonify({'objects': []}), 200
 
 
 @generator.response(status_code=201, schema={'id': 10, 'name': 'test_object'})
@@ -87,24 +98,16 @@ class AppTestBase(TestCase):
         super(AppTestBase, self).tearDown()
 
     def test(self):
-
-        generated_path = os.path.join(
-            pathlib.Path(__file__).parent.absolute()
+        generated_swagger_path = os.path.join(
+            pathlib.Path(__file__).parent.parent.absolute()
         ) + '/resources/generated.yaml'
-
-        reference_path = os.path.join(
-            pathlib.Path(__file__).parent.absolute()
-        ) + '/resources/reference_version_three.yaml'
-
-        generator.generate_swagger(
-            self.app, generated_path
-        )
-
-        generated_yaml = yaml_as_dict(generated_path)
-        reference_yaml = yaml_as_dict(reference_path)
-
+        correct_swagger_path = os.path.join(
+            pathlib.Path(__file__).parent.parent.absolute()
+        ) + '/resources/swagger_version_three/marshmallow_swagger.yaml'
+        generator.generate_swagger(self.app, generated_swagger_path)
+        generated_yaml = yaml_as_dict(generated_swagger_path)
+        reference_yaml = yaml_as_dict(correct_swagger_path)
         reference_yaml['info'].pop('description')
         generated_yaml['info'].pop('description')
-
         difference = DeepDiff(generated_yaml, reference_yaml, ignore_order=True)
         self.assertEqual({}, difference)
